routing / and /health to same checkalive method
