+++ weight/requirements_web.txt
@@ -1,0 +1,5 @@
+flask
+pandas
+mysql-connector-python
+flask-cors
+sqlalchemy
