--- weight/sandbox/containers1.csv
@@ -1,37 +1,0 @@
-"id","kg"
-C-35434,296
-C-73281,273
-C-35537,292
-C-49036,272
-C-85957,274
-C-57132,306
-C-80015,285
-C-40162,255
-C-66667,238
-C-65481,306
-C-65816,270
-C-38068,267
-C-36882,286
-C-38559,253
-C-83754,247
-C-40277,307
-C-55516,260
-C-45237,301
-C-69828,269
-C-44997,250
-C-52273,308
-C-63478,245
-C-42418,286
-C-86865,299
-C-38552,266
-C-81185,242
-C-71151,300
-C-78131,273
-C-61969,289
-C-82193,308
-C-85358,259
-C-47634,285
-C-83570,278
-C-45628,288
-C-70986,251
-C-54804,297
