--- weight/sandbox/containers2.csv
@@ -1,22 +1,0 @@
-"id","lbs"
-K-8263,666
-K-5269,666
-K-7943,644
-K-5355,642
-K-8120,657
-K-4987,653
-K-5867,666
-K-3963,541
-K-8298,554
-K-4722,530
-K-7947,589
-K-6176,523
-K-3690,682
-K-7488,633
-K-8128,528
-K-4743,572
-K-6350,550
-K-7425,677
-K-7714,525
-K-4109,587
-K-5768,569
