--- weight/sandbox/sql scripting tests/grantpriv.sql
@@ -1,2 +1,0 @@
-GRANT ALL PRIVILEGES ON *.* TO 'root'@'%' IDENTIFIED BY 'pass';
--- GRANT ALL PRIVILEGES ON \*.\* TO 'user'@'%' WITH GRANT OPTION;
