--- weight/sandbox/sql scripting tests/init.sql
@@ -1,3 +1,0 @@
-CREATE USER 'web'@'%' IDENTIFIED BY 'pass';
-GRANT ALL PRIVILEGES ON *.* TO 'web'@'%' IDENTIFIED BY 'pass';
-GRANT ALL PRIVILEGES ON *.* TO 'web'@'localhost';
